+++ test/utest/inference/embedding_feature_combiner_test.cpp
@@ -91,7 +91,7 @@ void embedding_feature_combine_test(int batch_size, int slot_num, int embedding_
   std::shared_ptr<IDataSimulator<int>> ldata_sim;
   ldata_sim.reset(new IntUniformDataSimulator<int>(0, max_nnz));
   h_row_ptrs[0] = 0;
-  for (int i = 1; i < (int)row_ptrs_size; i++) {
+  for (size_t i = 1; i < row_ptrs_size; i++) {
     h_row_ptrs[i] = (h_row_ptrs[i - 1] + ldata_sim->get_num());
   }
